move 2 chiral check fxns to chiral_utils.py
